Just a small update
This is what the code looks like right now. Still not working, but I’ve
added some list splits that should be needed. The main problem is that
I get a undefined procedure parse/3, but that there is a parse/5, even
when I move the list splitting to the right hand side.